Improve instructions for emulating space2ctrl in readme
In my old approach I bound space to a brand new keysym. This works well for
most programs but not for emacs! The problem is that emacs tries to warn
you when you press unbound keys, which causes slowdowns and odd behaviour.

The solution is to instead bind space to some unused modifier key that will
be recognised and ignored by emacs. I've found that "Hyper_L" works nicely.